Move v1 formatter into formatters folder
